+++ CHANGELOG.md
@@ -8,6 +8,14 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds the `role` prop to the `Text` component.
+
+### Updates
+
+- Updates the ARIA attributes for the `overline` and `subtitle` elements in the `Heading` component.
+
 ## 1.7.0 (July 20, 2023)
 
 ### Adds